+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enigma: Fix finding of a private key when decrypting a message using GnuPG v2.3
 - Fix TinyMCE localization installation (#9266)
 - Makefile: Use phpDocumentor v3.4 for the Framework docs (#9313)
+- Fix command injection via crafted im_convert_path/im_identify_path on Windows
 
 ## Release 1.5.6